# Minor performance improvements when serving static files