<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes but try to be as concise as possible -->
This PR improves slightly the performance of SWS when serving static files.

**Linux x86_64 (dynamic binary)**

Intel Core i7 8GB
Linux 6.7.9-arch1-1
Rust 1.76.0

```sh
# RAM ~6.2/6.4MB
wrk -c100 -t4 -d1m --latency http://localhost/index.html
# Running 1m test @ http://localhost/index.html
#   4 threads and 100 connections
#   Thread Stats   Avg      Stdev     Max   +/- Stdev
#     Latency     1.97ms    1.21ms  21.90ms   73.30%
#     Req/Sec    12.59k     1.17k   16.32k    75.21%
#   Latency Distribution
#      50%    1.72ms
#      75%    2.54ms
#      90%    3.60ms
#      99%    5.84ms
#   3008221 requests in 1.00m, 2.15GB read
# Requests/sec:  50073.47
# Transfer/sec:     36.63MB
```

**v2.28.0**

```sh
# RAM ~6.6/6.8MB
wrk -c100 -t4 -d1m --latency http://localhost/index.html
# Running 1m test @ http://localhost/index.html
#   4 threads and 100 connections
#   Thread Stats   Avg      Stdev     Max   +/- Stdev
#     Latency     2.02ms    1.24ms  28.34ms   74.09%
#     Req/Sec    12.31k     1.26k   21.35k    73.20%
#   Latency Distribution
#      50%    1.77ms
#      75%    2.60ms
#      90%    3.67ms
#      99%    6.04ms
#   2942017 requests in 1.00m, 2.10GB read
# Requests/sec:  48969.73
# Transfer/sec:     35.82MB
```


## Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment and the tests you ran to -->
<!--- See how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):